Last configuration JWT 1 - add line in file Kernel.php      'jwt.auth' => \Tymon\JWTAuth\Middleware\GetUserFromToken::class      middleware 2 - use jwt  pour authentification 3 - link RestFull DELETE http://localhost:8000/api/v1/meeting/registration/4?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOjUsImlzcyI6Imh0dHA6XC9cL2xvY2FsaG9zdDo4MDAwXC9hcGlcL3YxXC91c2VyXC9zaWduaW4iLCJpYXQiOjE0Njk5MTg0MDIsImV4cCI6MTQ2OTkyMjAwMiwibmJmIjoxNDY5OTE4NDAyLCJqdGkiOiI4YmQwZTA0NThjZDYwNTIyM2YyMjNkNzZmNmFkZTEzNCJ9.znQunX5ZAq31kqkzSDegdqic9gFrMKwYC4gLCw4m70Y pour désinscrire avec le tokan parameter

+++ app/Http/Controllers/MeetingController.php
@@ -7,12 +7,15 @@ use Carbon\Carbon;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
+use JWTAuth;
 
 class MeetingController extends Controller
 {
     public function __construct()
     {
-        // $this->middleware('name');
+         $this->middleware('jwt.auth',['only' => [
+             'update', 'store', 'destroy'
+         ]]);
     }
 
     /**
@@ -49,13 +52,16 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(!$user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg'=>'User not found'],404);
+        }
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'title' => $title,
@@ -114,31 +120,23 @@ class MeetingController extends Controller
      */
     public function update(Request $request, $id)
     {
+
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
-
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(!$user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg'=>'User not found'],404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
-        $meeting = [
-            'title' => $title,
-            'description' => $description,
-            'time' => $time,
-            'user_id' => $user_id,
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
+        $user_id = $user->id;
         $meeting = Meeting::with('users')->findOrFail($id);
         if (!$meeting->users()->where('users.id', $user_id)->first()) {
-
             return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
         }
         $meeting->time = Carbon::createFromFormat('YmdHie', $time);
@@ -168,6 +166,13 @@ class MeetingController extends Controller
     public function destroy($id)
     {
        $meeting = Meeting::findOrFail($id);
+
+        if(!$user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg'=>'User not found'],404);
+        }
+        if (!$meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
+        }
        $users =  $meeting->users;
        $meeting->users()->detach();
         if(!$meeting->delete()){
